examples/auth/requirements.txt
starlette-admin
sqlalchemy
itsdangerous
uvicorn
